backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
werkzeug==3.0.1

pytest==7.4.3
pytest-cov==4.1.0
allure-pytest==2.13.2
requests==2.31.0